Move stage to demo cmake subdir
