+++ src/engine/demo/BlurStage.h
@@ -11,9 +11,9 @@
 #include <core/light/DirectionalLight.h>
 #include <core/light/SpotLight.h>
 #include <core/effects/Bloom.h>
-#include "UIStage.h"
-#include "UIFrameDecorator.h"
-#include "UIButton.h"
+#include "gui/UIStage.h"
+#include "gui/UIFrameDecorator.h"
+#include "gui/UIButton.h"
 
 class BlurStage : public UIStage {
 private:
