+++ src/engine/demo/DitheringStage.cpp
