+++ src/engine/demo/DitheringStage.h
@@ -1,38 +1,38 @@
-#ifndef LIGHTSTAGE_H
-#define LIGHTSTAGE_H
+#ifndef DITHERINGSTAGE_H
+#define DITHERINGSTAGE_H
 
 #include <memory>
+#include <core/light/Materials.h>
 #include <core/models/ModelRenderer.h>
 #include <core/particles/ParticleGenerator.h>
 #include <core/particles/ParticleRenderer.h>
 #include <core/models/Model.h>
 #include <core/light/PointLight.h>
 #include <core/light/DirectionalLight.h>
+#include <core/algorithm/palette/MedianCut.h>
 #include <core/light/SpotLight.h>
-#include "UIStage.h"
-#include "UIFrameDecorator.h"
-#include "UIButton.h"
+#include "gui/UIStage.h"
+#include "gui/UIFrameDecorator.h"
+#include "gui/UIButton.h"
 
-class LightStage : public UIStage {
+class DitheringStage : public UIStage {
 private:
-    float a = 0;
+    float a = 0, b = 0;
+    GLuint textureID;
+    int size = 0;
+    std::vector<fVec3> colors, palette;
+
     Model model, plane;
     std::shared_ptr<ShaderProgram> program;
     std::shared_ptr<UIComponent> temp;
     PointLight pointLight;
     DirectionalLight dirLight;
-    SpotLight spotLight;
-    bool dir, point, spot;
 
-    void light(ShaderProgram *program, std::string name, PointLight pointLight);
-    void light(ShaderProgram *program, std::string name, DirectionalLight dirLight);
-    void light(ShaderProgram *program, std::string name, SpotLight spotLight);
-
-    static std::shared_ptr<LightStage> instance;
-    LightStage();
+    static std::shared_ptr<DitheringStage> instance;
+    DitheringStage();
 
 public:
-    static const std::shared_ptr<LightStage> &getInstance();
+    static const std::shared_ptr<DitheringStage> &getInstance();
 
     void setXxx(std::function<void ()> fun) {
         std::dynamic_pointer_cast<UIButton>(temp)->addClickCallback(fun);
@@ -42,4 +42,4 @@ public:
     void renderContent(Camera camera, double dt) override;
 };
 
-#endif // LIGHTSTAGE_H
+#endif // DITHERINGSTAGE_H
