+++ src/engine/demo/LightStage.cpp
