+++ src/engine/demo/LightStage.h
@@ -1,8 +1,7 @@
-#ifndef TEXTURESTAGE_H
-#define TEXTURESTAGE_H
+#ifndef LIGHTSTAGE_H
+#define LIGHTSTAGE_H
 
 #include <memory>
-#include <core/light/Materials.h>
 #include <core/models/ModelRenderer.h>
 #include <core/particles/ParticleGenerator.h>
 #include <core/particles/ParticleRenderer.h>
@@ -10,30 +9,30 @@
 #include <core/light/PointLight.h>
 #include <core/light/DirectionalLight.h>
 #include <core/light/SpotLight.h>
-#include "UIStage.h"
-#include "UIFrameDecorator.h"
-#include "UIButton.h"
+#include "gui/UIStage.h"
+#include "gui/UIFrameDecorator.h"
+#include "gui/UIButton.h"
 
-class TextureStage : public UIStage {
+class LightStage : public UIStage {
 private:
-    float a = 0, b = 0;
-    bool enableNormalMap = false;
-    bool enableSpecularMap = false;
-
+    float a = 0;
     Model model, plane;
     std::shared_ptr<ShaderProgram> program;
     std::shared_ptr<UIComponent> temp;
     PointLight pointLight;
     DirectionalLight dirLight;
+    SpotLight spotLight;
+    bool dir, point, spot;
 
     void light(ShaderProgram *program, std::string name, PointLight pointLight);
     void light(ShaderProgram *program, std::string name, DirectionalLight dirLight);
+    void light(ShaderProgram *program, std::string name, SpotLight spotLight);
 
-    static std::shared_ptr<TextureStage> instance;
-    TextureStage();
+    static std::shared_ptr<LightStage> instance;
+    LightStage();
 
 public:
-    static const std::shared_ptr<TextureStage> &getInstance();
+    static const std::shared_ptr<LightStage> &getInstance();
 
     void setXxx(std::function<void ()> fun) {
         std::dynamic_pointer_cast<UIButton>(temp)->addClickCallback(fun);
@@ -43,4 +42,4 @@ public:
     void renderContent(Camera camera, double dt) override;
 };
 
-#endif // TEXTURESTAGE_H
+#endif // LIGHTSTAGE_H
