+++ src/engine/demo/ShadingStage.cpp
