+++ src/engine/demo/TextureStage.h
@@ -1,5 +1,5 @@
-#ifndef DITHERINGSTAGE_H
-#define DITHERINGSTAGE_H
+#ifndef TEXTURESTAGE_H
+#define TEXTURESTAGE_H
 
 #include <memory>
 #include <core/light/Materials.h>
@@ -9,18 +9,16 @@
 #include <core/models/Model.h>
 #include <core/light/PointLight.h>
 #include <core/light/DirectionalLight.h>
-#include <core/algorithm/palette/MedianCut.h>
 #include <core/light/SpotLight.h>
-#include "UIStage.h"
-#include "UIFrameDecorator.h"
-#include "UIButton.h"
+#include "gui/UIStage.h"
+#include "gui/UIFrameDecorator.h"
+#include "gui/UIButton.h"
 
-class DitheringStage : public UIStage {
+class TextureStage : public UIStage {
 private:
     float a = 0, b = 0;
-    GLuint textureID;
-    int size = 0;
-    std::vector<fVec3> colors, palette;
+    bool enableNormalMap = false;
+    bool enableSpecularMap = false;
 
     Model model, plane;
     std::shared_ptr<ShaderProgram> program;
@@ -28,11 +26,14 @@ private:
     PointLight pointLight;
     DirectionalLight dirLight;
 
-    static std::shared_ptr<DitheringStage> instance;
-    DitheringStage();
+    void light(ShaderProgram *program, std::string name, PointLight pointLight);
+    void light(ShaderProgram *program, std::string name, DirectionalLight dirLight);
+
+    static std::shared_ptr<TextureStage> instance;
+    TextureStage();
 
 public:
-    static const std::shared_ptr<DitheringStage> &getInstance();
+    static const std::shared_ptr<TextureStage> &getInstance();
 
     void setXxx(std::function<void ()> fun) {
         std::dynamic_pointer_cast<UIButton>(temp)->addClickCallback(fun);
@@ -42,4 +43,4 @@ public:
     void renderContent(Camera camera, double dt) override;
 };
 
-#endif // DITHERINGSTAGE_H
+#endif // TEXTURESTAGE_H
